Remove network observer module
